chore: update types system
