+++ src/stores/themeDataSlice/config.ts
@@ -1,5 +1,3 @@
-import { CustomTheme } from '.';
-
 export const variableVals = [
   '--background',
   '--foreground',
