+++ src/stores/themeDataSlice/index.ts
@@ -1,4 +1,3 @@
 export * from './themeDataSlice';
-export * from './types';
 export * from './config';
 export * from './actions';
